+++ package.json
@@ -13,6 +13,7 @@
     "react-bootstrap": "^1.5.1",
     "react-dom": "^17.0.1",
     "react-redux": "^7.2.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "sass": "^1.32.8",
     "typescript": "^4.2.2",
