+++ src/App.tsx
@@ -1,9 +1,18 @@
-import MainPage from './components/MainPage/MainPage';
+import { Container, Row } from 'react-bootstrap';
+import Footer from './components/Footer/Footer';
+import Header from './components/Header/Header';
+import Routes from './Routes';
 
 const App: React.FC = () => {
   return (
     <div className="App">
-      <MainPage />
+      <Header />
+      <Container>
+        <Row>
+          <Routes />
+        </Row>
+      </Container>
+      <Footer />
     </div>
   );
 };
