+++ src/components/Footer/Footer.tsx
@@ -1,4 +1,4 @@
-const Footer = () => {
+const Footer: React.FC = () => {
   return <footer>Footer</footer>;
 };
 export default Footer;
