+++ src/components/Header/Header.tsx
@@ -1,4 +1,4 @@
-const Header = () => {
+const Header: React.FC = () => {
   return <header>Header</header>;
 };
 export default Header;
