+++ src/pages/Country/Country.tsx
@@ -1,4 +1,4 @@
-const Country = () => {
+const Country = (): JSX.Element => {
   return <div>Country</div>;
 };
 export default Country;
